feat(parser): Refactor Markdown handling and improve structure
- Introduced a new `MarkdownPatterns` module to centralize regular expression patterns for various Markdown elements, enhancing maintainability and readability.
- Updated the `Parser` class to utilize the new `MarkdownPatterns` for improved Markdown element handling.
- Removed redundant metadata from the `TestCaseElement` interface related to Markdown, simplifying the structure and focusing on content type.
- Refactored the `formatMarkdownElement` method to apply Markdown formatting through a dedicated function, streamlining the parsing process.

These changes enhance the parser's efficiency in handling Markdown content while maintaining a clean and modular codebase.
