+++ src/parser/tokenizer.ts
@@ -1,4 +1,5 @@
 import { ParseError } from './errors.js';
+import { MarkdownPatterns } from './markdown-handlers.js';
 
 /**
  * Represents an element in the test case text
@@ -7,34 +8,12 @@ export interface TestCaseElement {
   /**
    * The type of the element
    */
-  type: 'role' | 'content' | 'tool_use' | 'tool_response' | 'args' | 'markdown';
+  type: 'role' | 'content' | 'tool_use' | 'tool_response' | 'args';
 
   /**
    * The value of the element
    */
   value: string;
-
-  /**
-   * Optional metadata for the element
-   */
-  metadata?: {
-    markdownType?:
-      | 'header'
-      | 'list'
-      | 'code'
-      | 'blockquote'
-      | 'table'
-      | 'horizontal_rule'
-      | 'task_list'
-      | 'code_block_start'
-      | 'code_block_end';
-    level?: number; // For headers (1-6)
-    listType?: 'ordered' | 'unordered' | 'task';
-    isNested?: boolean;
-    language?: string;
-    marker?: string;
-    indentLevel?: number;
-  };
 }
 
 /**
@@ -63,15 +42,15 @@ export class TestCaseParser {
     TOOL_RESPONSE: /^\s*tool response:/i,
     ARGS: /^args:/i,
     // Markdown patterns
-    MARKDOWN_HEADER: /^(#{1,6})\s+(.+)$/,
-    MARKDOWN_LIST: /^(\s*)([*+-]|\d+\.)\s+(.+)$/,
-    MARKDOWN_TASK_LIST: /^(\s*)[*+-]\s+\[([ xX])\]\s+(.+)$/,
-    MARKDOWN_CODE_BLOCK: /^```([\w-]*)\n([\s\S]*?)```$/,
-    MARKDOWN_INLINE_CODE: /`([^`]+)`/,
-    MARKDOWN_BOLD: /\*\*([^*]+)\*\*/,
-    MARKDOWN_ITALIC: /\*([^*]+)\*/,
-    MARKDOWN_BLOCKQUOTE: /^>\s+(.+)$/,
-    MARKDOWN_HORIZONTAL_RULE: /^([*\-_])\s*\1\s*\1$/,
+    MARKDOWN_HEADER: MarkdownPatterns.MARKDOWN_HEADER,
+    MARKDOWN_LIST: MarkdownPatterns.MARKDOWN_LIST,
+    MARKDOWN_TASK_LIST: MarkdownPatterns.MARKDOWN_TASK_LIST,
+    MARKDOWN_CODE_BLOCK: MarkdownPatterns.MARKDOWN_CODE_BLOCK,
+    MARKDOWN_INLINE_CODE: MarkdownPatterns.MARKDOWN_INLINE_CODE,
+    MARKDOWN_BOLD: MarkdownPatterns.MARKDOWN_BOLD,
+    MARKDOWN_ITALIC: MarkdownPatterns.MARKDOWN_ITALIC,
+    MARKDOWN_BLOCKQUOTE: MarkdownPatterns.MARKDOWN_BLOCKQUOTE,
+    MARKDOWN_HORIZONTAL_RULE: MarkdownPatterns.MARKDOWN_HORIZONTAL_RULE,
   };
 
   /**
@@ -152,34 +131,17 @@ export class TestCaseParser {
       pattern: TestCaseParser.PATTERNS.MARKDOWN_HEADER,
       handle: (line: string, elements: TestCaseElement[]) => {
         elements.push({
-          type: 'markdown',
+          type: 'content',
           value: line,
-          metadata: {
-            markdownType: 'header',
-            level: (line.match(/^#+/) || [''])[0].length,
-          },
         });
       },
     },
     {
       pattern: TestCaseParser.PATTERNS.MARKDOWN_LIST,
       handle: (line: string, elements: TestCaseElement[]) => {
-        const match = line.match(TestCaseParser.PATTERNS.MARKDOWN_LIST);
-        if (!match) return;
-        const indent = match[1];
-        const marker = match[2];
-        const listType = /^\d+\.$/.test(marker) ? 'ordered' : 'unordered';
-        const isNested = indent.length > 0;
         elements.push({
-          type: 'markdown',
+          type: 'content',
           value: line,
-          metadata: {
-            markdownType: 'list',
-            listType,
-            isNested,
-            marker,
-            indentLevel: indent.length,
-          },
         });
       },
     },
@@ -187,11 +149,8 @@ export class TestCaseParser {
       pattern: TestCaseParser.PATTERNS.MARKDOWN_TASK_LIST,
       handle: (line: string, elements: TestCaseElement[]) => {
         elements.push({
-          type: 'markdown',
+          type: 'content',
           value: line,
-          metadata: {
-            markdownType: 'task_list',
-          },
         });
       },
     },
@@ -199,11 +158,8 @@ export class TestCaseParser {
       pattern: TestCaseParser.PATTERNS.MARKDOWN_BLOCKQUOTE,
       handle: (line: string, elements: TestCaseElement[]) => {
         elements.push({
-          type: 'markdown',
+          type: 'content',
           value: line,
-          metadata: {
-            markdownType: 'blockquote',
-          },
         });
       },
     },
@@ -211,11 +167,8 @@ export class TestCaseParser {
       pattern: TestCaseParser.PATTERNS.MARKDOWN_HORIZONTAL_RULE,
       handle: (line: string, elements: TestCaseElement[]) => {
         elements.push({
-          type: 'markdown',
+          type: 'content',
           value: line,
-          metadata: {
-            markdownType: 'horizontal_rule',
-          },
         });
       },
     },
@@ -243,22 +196,16 @@ export class TestCaseParser {
         if (!inCodeBlock) {
           // Start of code block
           inCodeBlock = true;
-          codeBlockLanguage = line.slice(3);
           codeBlockLines = [line];
         } else {
           // End of code block
           inCodeBlock = false;
           codeBlockLines.push(line);
           elements.push({
-            type: 'markdown',
+            type: 'content',
             value: codeBlockLines.join('\n'),
-            metadata: {
-              markdownType: 'code',
-              language: codeBlockLanguage.trim(),
-            },
           });
           codeBlockLines = [];
-          codeBlockLanguage = '';
         }
         continue;
       }
@@ -346,16 +293,3 @@ export class TestCaseParser {
     }
   }
 }
-
-export const MarkdownPatterns = {
-  // Markdown patterns
-  MARKDOWN_HEADER: /^(#{1,6})\s+(.+)$/,
-  MARKDOWN_LIST: /^(\s*)([*+-]|\d+\.)\s+(.+)$/,
-  MARKDOWN_TASK_LIST: /^(\s*)[*+-]\s+\[([ xX])\]\s+(.+)$/,
-  MARKDOWN_CODE_BLOCK: /^```([\w-]*)\n([\s\S]*?)```$/,
-  MARKDOWN_INLINE_CODE: /`([^`]+)`/,
-  MARKDOWN_BOLD: /\*\*([^*]+)\*\*/,
-  MARKDOWN_ITALIC: /\*([^*]+)\*/,
-  MARKDOWN_BLOCKQUOTE: /^>\s+(.+)$/,
-  MARKDOWN_HORIZONTAL_RULE: /^([*\-_])\s*\1\s*\1$/,
-};
